+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/block/processor/CompactBlockProcessor.kt
@@ -503,19 +503,6 @@ class CompactBlockProcessor internal constructor(
                         // Invalidate transaction data
                         checkTransactions(transactionStorage = repository)
                     }
-                    SyncingResult.FetchUtxos -> {
-                        Twig.info { "Triggering UTXOs fetching from: ${fullyScannedHeight.value}" }
-                        if (fullyScannedHeight.value == null) {
-                            Twig.info { "Postponing UTXOs fetching because fullyScannedHeight is null" }
-                        } else {
-                            backend.getAccounts().forEach {
-                                refreshUtxos(
-                                    account = it,
-                                    startHeight = fullyScannedHeight.value!!
-                                )
-                            }
-                        }
-                    }
                     is SyncingResult.Failure -> {
                         syncingResult = batchSyncProgress.resultState
                         return@collect
@@ -614,20 +601,6 @@ class CompactBlockProcessor internal constructor(
                         Twig.info { "Triggering transaction refresh now" }
                         // Invalidate transaction data and return the common batch syncing success result to the caller
                         checkTransactions(transactionStorage = repository)
-                        SyncingResult.AllSuccess
-                    }
-                    SyncingResult.FetchUtxos -> {
-                        Twig.info { "Triggering UTXOs fetching from: ${fullyScannedHeight.value}" }
-                        if (fullyScannedHeight.value == null) {
-                            Twig.info { "Postponing UTXOs fetching because fullyScannedHeight is null" }
-                        } else {
-                            backend.getAccounts().forEach {
-                                refreshUtxos(
-                                    account = it,
-                                    startHeight = fullyScannedHeight.value!!
-                                )
-                            }
-                        }
                         SyncingResult.AllSuccess
                     }
                     is SyncingResult.Failure -> {
@@ -1484,15 +1457,6 @@ class CompactBlockProcessor internal constructor(
                         }.buffer(1)
                         .map { downloadStageResult ->
                             Twig.debug { "Download stage done with result: $downloadStageResult" }
-
-                            // Triggering UTXOs fetch operation
-                            emit(
-                                BatchSyncProgress(
-                                    order = downloadStageResult.batch.order,
-                                    range = downloadStageResult.batch.range,
-                                    resultState = SyncingResult.FetchUtxos
-                                )
-                            )
 
                             if (downloadStageResult.stageResult !is SyncingResult.DownloadSuccess) {
                                 // In case of any failure, we just propagate the result