+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/block/processor/model/SyncingResult.kt
@@ -61,8 +61,6 @@ internal sealed class SyncingResult {
 
     data object EnhanceSuccess : SyncingResult()
 
-    data object FetchUtxos : SyncingResult()
-
     data class EnhanceFailed(
         override val failedAtHeight: BlockHeight? = null,
         override val exception: CompactBlockProcessorException