+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- The note commitment tree bug has been resolved using a new internal `Backend.fixWitnesses()` API 
+
 ### Changed
 - Dependency update:
   - Gradle 8.13
@@ -13,6 +16,10 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Kotlin 2.1.10
   - Bip39 1.0.9
   - Other dependencies update
+- Migrated to `zcash_client_backend 0.18.0`, `zcash_client_sqlite 0.16.0`
+- Added support for gap-limit-based discovery of transparent wallet addresses.
+- The internal `fetch-utxos` logic is now triggered only in every `init` and `complete` block sync phases, and it 
+  fetches UTXOs from height 0 to support the Ledger funds rescue requirement.
 - Checkpoints update
 
 ## [2.2.8] - 2025-03-03